rktio: normalize 'replace and 'truncate/replace behavior
For `open-output-file`:

Move 'replace mode handling out of rktio and into the client.
Formerly, on Windows, 'replace mode was just 'truncate with a fallback
for permission problems, so this change makes it delete an existing
file and replace it, which is more consistent with how 'replace has
always worked on Unix.

In 'truncate/replace mode, if a truncating open fails due to a
permission error and the file exists, then try again as a 'replace.
That's how it worked on Windows before, and now it's how Unix works.

+++ racket/src/rktio/demo.c
@@ -704,7 +704,7 @@ int main(int argc, char **argv)
     rktio_fd_t **pipe_fds;
 
     if (stress && verbose)
-      printf(" iter %d\n", i);
+      printf(" iter %d\n", (int)i);
 
     pipe_fds = rktio_make_pipe(rktio, 0);
     check_valid(pipe_fds);
@@ -756,7 +756,7 @@ int main(int argc, char **argv)
     rktio_listener_t *lnr;
 
     if (stress && verbose)
-      printf(" iter %d\n", i);
+      printf(" iter %d\n", (int)i);
 
     check_many_lookup(rktio);
